Large update
Separated character_class.js into two files, 'character_class_info.js' and 'character_class_display.js'. The former focuses on functions that affect the data of the Character Class, while the latter focuses on functions that display that data on the page. Also refactored dice_roller.js to be able to record the Verbose of a dice roll, as well as procedurally generate the Manual Dice Roller from Javascript to reduce the size within the HTML file. Decided to stay away from Collapsible Sections at the moment, and migrated to a Tab-based page for easier use, which involves having multiple divs on the page at once and hiding/showing them depending on the button pressed in the new Navbar. Also added files to support a new custom Dropdown version of the Select box, that is actually created using a button and lists, to allow for selecting from Nested lists (e.g. the Lore skill). Finally, removed outstanding data files that just weren't being used at all, and started tidying up the functions for the data accessors (which are now just their own files in the /data folder) to make sure that when accessing the data, results were only being passed by value and not by reference (this caused a few bugs).