Tidied character_page.html, added DataHandler class
Added DataHandler class, which loads required information (such as skill info) from files hosted online, instead of loading them through custom Javascript objects. Also added Advantages and Disadvantages sections to the Character tab, and removed the integrated-dice-roller (as opposed to the Standalone Dice Roller) as I am unlikely to use it at this point in time. With things like Advantages and Void points, it may be better off to just provide all the information on-screen, and let people use the manual dice roller.