Fixed validation rules for Advantages
Also updated Advantages.json file so that all discounts are assumed to be nothing by default, allow_multiple is assumed to be false by default, and all inputs are assumed to be required by default. Any excess values from this (e.g. stating that an input was required, or that there was no discount) were stripped.