DataHandler is now a global (window) variable, refactored Advantages, added placeholder info
DataHandler is now a global variable in window.dh, so it can be accessed more easily by any object, instead of continually passed from one to another even though we only ever need one instance of it. Code for creating Advantages has been refactored slightly to condense and streamline it. Added in some placeholders for additional unimportant information (like Family and Physical Appearance) and moved character name to the top of the page.