Added in recommended fixes from Feedback, and then some
BIG CHANGES:

Added PageOptions class, with an access button in the top right corner.
Allows the user to modify certain page options outside of the
CharacterCreator, and will
turning Autosave on and off, and toggling Dark Mode.

Added a Dark Mode for the page using an alternate style sheet (.css).
Still a work in progress for now, but should be easy to do in future
using CSS variables (--variable-name-here).

Added a Save and Load feature to the CharacterCreator. Also added an
Autosave feature, which can be controlled through the use of the
PageOptions. Existing character will be loaded on page load, if found.

Added in displays for the character's Experience, Insight, and Insight
Rank, which are dynamically calculated on the fly.
Changed other displays (e.g. Gear) to be retrieved from DataHandler and
displayed dynamically as well, instead of storing data that we don't
need to.

'Different School' advantage has been removed from the Advantages.json
file, and is now built into the CharacterCreator.refresh_adv_display
method. Will need to find somewhere else to store it eventually. It is
automatically applied to a character if they have a School that is from
affecting their options.

FIXED: custom_dropdown menus for skill selection now hang correctly
underneath their corresponding buttons, by displaying in a Flexbox
div.

Add_Custom_Spell code has been moved to the CustomData object in
Data_handler.js, for potential future use.

Removed reference to display_functions.js from character_creation.js, as
it was no longer used.
Added a Select so users can pick character's Affinity if allowed (pretty
much ONLY Isawa Shugenja).

Raises and Special text in Spells now only display if they are present
for that spell.

School Select informing players of this fact, when the condition is met.

Added customised Delete buttons for the Advantages and Spells tables, to
aid in removing thos
'no-bg' class allows cells to 'remove' the background color of their row
by setting it to the main background color.
Added 'error' and 'note' classes.
Added various stylings for Delete Buttons and Table cells.

Selects now display in italic when they have a default option set (have
to specify this manually in the code). Does not affect options.

Custom Dropdowns can also be made to be Italic when necessary, without
affecting options.