Updated Collapsible Divs
Changed classname 'Hidden' to 'Closed', as it is more accurately representative. Titles can now include more than just text (e.g. Buttons). Changed div content offset from Margin to Padding, so that it doesn't affect any in-content measurements (subheadings were sticking out to the right). Added Subheadings as a class, which is a black title with a border box enclosing the contents.