Added ability to check of spells exceed spell choices
Spells can now be removed if they exceed the spell choices, spells of a certain element can no longer be selected if you have too many of that element already. Have not completed exception for Isawa Shugenja. All Spell Table is now created once, and the spells for each rank are updated each time it is refreshed, instead of the whole table being recreated. This allows the user to preserve which tabs are open in the table.