Added Charts to dice_roller
Added roll probability charts to dice_roller through CanvasJS. Added corresponding CSS and JS. Unlocked Roll History to reintegrate (not completed yet). Running the simulation required for the Charts is causing a lag on the Dice Roller, and will have to be moved to a Worker process to prevent this; Core dice rolling functions have been separated into a new object called DiceRollerBaseFunctions for this purpose.